CORE-split out view and dialog
